% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{exprByType}
\alias{exprByType}
\title{simple plot of expression values by structure/expression-based selection in IvyGAP}
\usage{
exprByType(sym, ...)
}
\arguments{
\item{sym}{a gene symbol found among `rownames(ivySE)`}

\item{\dots}{passed to plot, exclusive of ylab, xlab, axes}
}
\description{
simple plot of expression values by structure/expression-based selection in IvyGAP
}
\examples{
exprByType("MYC")
}
