% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tumorDetails}
\alias{tumorDetails}
\alias{subBlockDetails}
\alias{vocab}
\title{helper functions for data access}
\usage{
tumorDetails(se)

subBlockDetails(se)

vocab()
}
\arguments{
\item{se}{SummarizedExperiment instance, intended to work for ivySE in this package}
}
\description{
helper functions for data access
}
