% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgs.R
\name{designOverview}
\alias{designOverview}
\alias{nomenclat}
\title{render design overview}
\usage{
designOverview()

nomenclat()
}
\value{
a rastergrob grob
}
\description{
render design overview

render anatomic nomenclature
}
\examples{
designOverview()
}
