% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ish_stem_details}
\alias{ish_stem_details}
\title{details on image quantification results with ISH}
\format{
An object of class \code{data.frame} with 18778 rows and 29 columns.
}
\source{
\url{http://glioblastoma.alleninstitute.org/api/v2/gbm/gene_expression_details.csv}
}
\usage{
ish_stem_details
}
\description{
details on image quantification results with ISH
}
\examples{
data(ish_stem_details)
names(ish_stem_details)
}
\keyword{datasets}
